refactor(vscode): move node client options
